requirements.txt
absl-py==1.4.0
astunparse==1.6.3
conda==23.3.1
contourpy==1.1.0
fonttools==4.40.0
fsspec==2023.6.0
gast==0.4.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.9.0
joblib==1.2.0
jsonpointer==2.1
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
nltk==3.8.1
numpy==1.24.3
opencv-python==4.8.0.74
opt-einsum==3.3.0
patool==1.12
Pillow==10.0.0
protobuf==4.23.3
pyparsing==3.0.9
python-dateutil==2.8.2
pyunpack==0.3
regex==2022.10.31
scikit-learn==1.2.2
sklearn==0.0.post1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.1.0
tokenizers==0.13.3
transformers==4.30.2
typing_extensions==4.5.0
Werkzeug==2.3.6
wrapt==1.15.0